properly set language for all activities, also on screeen rotation
