+++ src/org/thoughtcrime/securesms/ConversationListActivity.java
@@ -28,7 +28,6 @@ import static org.thoughtcrime.securesms.util.RelayUtil.isRelayingMessageContent
 import static org.thoughtcrime.securesms.util.RelayUtil.resetRelayingMessageContent;
 
 import android.content.Intent;
-import android.content.res.Configuration;
 import android.net.Uri;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -63,7 +62,6 @@ import org.thoughtcrime.securesms.qr.QrActivity;
 import org.thoughtcrime.securesms.qr.QrCodeHandler;
 import org.thoughtcrime.securesms.recipients.Recipient;
 import org.thoughtcrime.securesms.search.SearchFragment;
-import org.thoughtcrime.securesms.util.DynamicLanguage;
 import org.thoughtcrime.securesms.util.DynamicNoActionBarTheme;
 import org.thoughtcrime.securesms.util.DynamicTheme;
 import org.thoughtcrime.securesms.util.Prefs;
@@ -80,9 +78,6 @@ public class ConversationListActivity extends PassphraseRequiredActionBarActivit
   public static final String CLEAR_NOTIFICATIONS = "clear_notifications";
   public static final String ACCOUNT_ID_EXTRA = "account_id";
 
-  private final DynamicTheme    dynamicTheme    = new DynamicNoActionBarTheme();
-  private final DynamicLanguage dynamicLanguage = new DynamicLanguage();
-
   private ConversationListFragment conversationListFragment;
   public TextView                  title;
   private AvatarImageView          selfAvatar;
@@ -93,8 +88,8 @@ public class ConversationListActivity extends PassphraseRequiredActionBarActivit
 
   @Override
   protected void onPreCreate() {
-    dynamicTheme.onCreate(this);
-    dynamicLanguage.onCreate(this);
+    dynamicTheme = new DynamicNoActionBarTheme();
+    super.onPreCreate();
   }
 
   @Override
@@ -252,18 +247,7 @@ public class ConversationListActivity extends PassphraseRequiredActionBarActivit
   @Override
   public void onResume() {
     super.onResume();
-    dynamicTheme.onResume(this);
-    dynamicLanguage.onResume(this);
-
     DirectShareUtil.triggerRefreshDirectShare(this);
-  }
-
-  @Override
-  public void onConfigurationChanged(Configuration newConfig) {
-    Log.i(TAG, "onConfigurationChanged(orientation=" + newConfig.orientation + ")");
-    super.onConfigurationChanged(newConfig);
-    // on orientation changes locale is reset in the context/activity so set locale as onCreate()
-    dynamicLanguage.onCreate(this);
   }
 
   @Override
