+++ src/org/thoughtcrime/securesms/PassphraseRequiredActionBarActivity.java
@@ -13,7 +13,6 @@ public abstract class PassphraseRequiredActionBarActivity extends BaseActionBarA
   @Override
   protected final void onCreate(Bundle savedInstanceState) {
     Log.w(TAG, "onCreate(" + savedInstanceState + ")");
-    onPreCreate();
 
     if (GenericForegroundService.isForegroundTaskStarted()) {
       // this does not prevent intent set by onNewIntent(),
@@ -37,6 +36,5 @@ public abstract class PassphraseRequiredActionBarActivity extends BaseActionBarA
     }
   }
 
-  protected void onPreCreate() {}
   protected void onCreate(Bundle savedInstanceState, boolean ready) {}
 }
